Modify updateActualImage to handle elilo-style args
elilo handles hypervisor args by putting a "--" separator on the
append line, like this:

    append="hyper args -- kernel args"

This patch modifies updateActualImage() to handle this situation,
including removing the hypervisor args plus separator when a multiboot
template is used to construct a non-multiboot entry.

Grubby builds with this patch, and the test suite doesn't report any
new regressions.

Signed-off-by: Aron Griffis <aron@hp.com>