make new-kernel-pkg add the necessary cmdline params for dracut
add a --add-dracut-args cmdline option to new-kernel-pkg, which will make it
add the necessary kernel cmdline params for i18n, keyboard and
plymouth when installing a dracut kernel.

This goes hand in hand with a kernel.spec patch which makes dracut using
kernels pass this option in their %post script.