grubby: compare rootdev using uuid instead of stat
Btrfs creates anonymous backing devices for its superblocks in order to deal
with the fact that it can handle multiple devices.  This means that if you stat
the mountpoint of a btrfs filesystem, you will get a different device than the
device you mounted with.  This causes problems for Grubby because it fails out
if the / device and the device that is currently mounted are not the same.
Instead of doing the stat check, open up /etc/mtab instead and look for the
device we used to mount the filesystem with and then get that device's uuid and
compare it against the device we think the root is to make sure everything is
OK.  I've tested this patch on a box with a btrfs root and now grubby does not
fail.

Signed-off-by: Josef Bacik <josef@redhat.com>