new-kernel-pkg fails to find U-Boot.
new-kernel-pkg fails to find U-Boot if the boot directory is not on a
separate partition.

This only affects the U-Boot image creation section of the new-kernel-pkg
script, for ARM based systems.  My initial changes were only tested on
specific ARM-OMAP systems that required a specific U-Boot partition.  Testing
on additional systems (ARM-Tegra) revealed this incorrect assumption.

The attached patch includes the following:

- Cosmetic: Correction - change uBoot to U-Boot in comments.
- Additions:
-- Recognize ARCH == armv7hl
-- Make U-Boot load/execute address variable (based on machine)
-- Bug: Correct invalid assumption that the U-Boot directory would always
   be on a separate partition.