grubby: Add --debug option to annotate "suitable template" failures
Today, when grubby fails with "no suitable template found" the user
has no idea why that might be.  These can fail for several reasons.
So add a --debug option to print out the line in grubby.c where it
failed, a text message describing the failure, and a copy of the
bootloader entry that failed, to aid investigation.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>