Support systems that do not have /boot as a separate filesystem
While testing grubby on systems with a root (/) filesystem that includes
the /boot directory, it was noticed that grubby displayed the prefix (/boot)
twice.  This patch prevents printing the prefix twice on kernel and initrd
lines.

This situation was found while testing the patch "Fix for grub on SuSE
systems: lba and boot information", so this fix is also covered by the
the test on patch "Fix for grub on SuSE systems: test lba and boot info".

Signed-off-by: Cleber Rosa <crosa@redhat.com>