Fix grub detection on SuSE systems: config file order
SuSE sytems ship with a /etc/grub.conf that is not your regular grub
configuration file, but instead a sequence of grub commands. Example:

setup --stage2=/boot/grub/stage2 --force-lba (hd0,1) (hd0,1)
quit

Because of that, the parsing of the config fails. So, try to first
use the grub config file at /boot/grub/menu.lst.

Signed-off-by: Cleber Rosa <crosa@redhat.com>