new-kernel-pkg: fix recursion with --mkinitrd (#832238)
commit eec2063c966df68a9e1947d9e9474df72540e3c6 ("Fix hard coded paths
so kernel's "make install" will DTRT.") accidentally introduced
recursion for --mkinitrd by having the mkinitrd function execute itself.

I ran in to this when inadvertantly using an old ~/bin/installkernel
script which was still using new-kernel-pkg --mkinitrd.

This patch avoids the recursion by renaming the shell wrapper doMkinitrd
(matching depmod which presumably had the same problem) so that it's not
executed instead of the mkinitrd that's in the path.

Signed-off-by: Zach Brown <zab@redhat.com>