Change the way the kernel load address is determined for ARM U-Boot.
The kernel load address was initially determined by using the
Hardware entry in /proc/cpuinfo to identify the platform.  This
involved a substring match that was inexact, and was based on the
host hardware, and not the package being installed.

This patch changes grubby to use the kernel flavor (last component
of the kernel-release) of the kernel package being installed to
determine the platform.  This approach also allows a kernel to
be installed from a host that differs from the target platform
when using tools like livemedia-creator.

Signed-off-by: David A. Marlin <dmarlin@redhat.com>