extlinux: Understand "default" properly
The configuration looks like the following:

  default Fedora (3.11.6-301.fc20.i686+PAE) 20 (Heisenbug)

  title Fedora (3.11.6-301.fc20.i686+PAE) 20 (Heisenbug)
  ...

Grubby skips over the default clause as it has more than one element. And even
if it did not, it would not match it against the title, since it handles titles
specially, concatenating the title, but not the default clause.

This commit adds special handling for extlinux, which causes it to parse
default and title in the same way.

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>