Strip the LT_END line from a new stanza before adding "missing" parts.
When we get --copy-default on a --add-kernel with a --initrd option, and
the default stanza lacks an initrd, we wind up in the fallback path
where we add the default after the fact.  When we do this, we later add
an LT_END line /if/ it doesn't exist.  But we've already got one, and
when we add the initrd it's after it.

So strip it out and re-mark it as needed so it'll be added at the end.

Resolves: rhbz#1099627
Related: rhbz#1096358

Signed-off-by: Peter Jones <pjones@redhat.com>