Don't go past the last element of indexVars in findEntryByPath
We add a chance of creating an infinite loop, because we
were reading memory past the last element of indexVars set to -1.

This issue was only apparent with -O2, probably because the way the
memory was initialized.

Signed-off-by: Yannick Brosseau <scientist@fb.com>
(cherry picked from commit e4a506d5a6aababd33571cc328fdd24dbdb894d1)

Resolves: rhbz#1238941