cleanup dtb handling to work in the supported usecases
add SHIPSDTB variable in the uboot defaults file that needs to be set to
yes for platforms like the calxeda highbank that ship a dtb in u-boot
that we want to use.  if the user defines a dtbfile in
/etc/sysconfig/uboot update the extlinux.conf with an fdt entry
otherwise update a fdtdir entry unless SHIPSDTB is set to yes.