grubby: properly handle mixed ' and " and nested quotes
The SLES12 grub2.cfg file on ppc64le by default contains a line like:

  submenu "Bootable snapshot #$snapshot_num" {
    menuentry "If OK, run 'snapper rollback $snapshot_num' reboot." { true; }
  }

On any grubby (tested with 8.40) invocation that updates the config
file, the combination of nested quotes and mixed quotes leads to a
generated file content like:

  submenu "Bootable snapshot #$snapshot_num" {
    menuentry 'If OK, run snapper rollback $snapshot_num' rollback $snapshot_num' and reboot." { true; }
  }

which includes both a change from " to ', but also improperly quoted
strings and trailing characters relative to the string. This actually
leads to a failure to boot from the disk by default when using grubby
(e.g., Autotest) on SLES12 ppc64le. Whether SLES12 should be adding an
entry like this by default or not is probably open to debate, but grubby
should be able to hand this input file.

To fix the issue, three changes were necessary:

1) grub2ExtractTitle needs to check that if the second element starts
with a quote, that the matching element found ends with the same
quote-type (' vs. ")

2) lineWrite needs to output the right kind of quote based upon if the
string to be outputted itself contains quotes. This is not currently
possible in the code, because quotes are stripped out normally by
readConfig, but with the change in 3), that only happens now for the
quotes that actually delineate a string.

3) readConfig needs to check that when it is extracting titles and
determining extras, it uses matching quotes.

With these changes, a simple grubby --set-default=SLES12 (for example),
now produces:

   submenu "Bootable snapshot #$snapshot_num" {
     menuentry "If OK, run 'snapper rollback $snapshot_num' and reboot." { true; }
   }

as expected.

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>