ppc64le sync grub.cfg changes to disk (#1212114)
After installing a new kernel, if a panic is induced, not all the kernel
pieces are on the disk (most importantly the grub.cfg changes).  This can
lead to a hung system on a reboot because the older kernel can not be found.

Address this by forcing all the changes (mainly the fs meta data) to disk
before finishing the kernel package installation.

Tested by 'yum install kernel-...; echo c > /proc/sysrq-trigger'.

Before, the machine would panic and on reboot be stuck without a grub.cfg
to read.  After, works as expected.

Related: rhbz#1212114