Every new boot entry is not default (#1285601)
Due to a logic issue grubby made the assumption that every time a new
boot entry gets added it should be the default. Due to this behavior, it
also masked logic issues that prevented the selection of the proper
index in cases where grubby was asked to dynamically pick a default boot
entry or keep the existing default boot entry.

Resolves: rhbz#1285601