Fix some test cases where the resulting default index is wrong
Fix two issues with tests where removing a kernel or adding a boot entry
at a user specified index had the wrong default index selected in the
expected output.

One of these issues is the --set-index test cases for grub 1, which
expect --make-default is implied where it should not be.

The third is a removal test, where the default index in the result data
is off by one when we've removed 2 entries keyed off of the kernel
filesystem path.

This change causes these test cases to fail.

Resolves: rhbz#1285601