Fix incorrect test case and --remove-args with a value.
Currently we have this test case:

  grubTest grub.3 updargs/g3.4 --update-kernel=ALL --remove-args="hdd=foobar"

This fails to notice that the actual argument in grub.3 is hdd=ide-scsi,
and removes it anyway, and the data in g3.4 supports that behavior.
This is clearly wrong, and so this patch introduces updargs/g3.5, which
leaves hdd=ide-scsi intact, and fixes the code so that it won't modify
the command line in that case.

Resolves: rhbz#1476273

Signed-off-by: Peter Jones <pjones@redhat.com>