Print default image even if isn't a suitable one
The grubby --default-kernel option only prints the default kernel if this
is a suitable one. That is if its associated kernel cmdline root param is
the same than the partition currently mounted as the filesystem root.

But the grubby --set-default option doesn't have that restriction, it is
able to set a kernel as the default even if its root is for a different
partition. So make the --default-kernel option to also print the kernel
in this case. Still check if is a suitable image so --debug can tell it.

Resolves: rhbz#1323842

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>